+++ packages/graphql/src/schema/mutation.ts
@@ -211,6 +211,13 @@ export const Mutation = builder.mutationType({
         },
       }),
 
+      publishScheduledPracticeQuizzes: t.boolean({
+        resolve(_, __, ctx) {
+          checkCronToken(ctx)
+          return PracticeQuizService.publishScheduledPracticeQuizzes(ctx)
+        },
+      }),
+
       createParticipantAccount: t.field({
         nullable: true,
         type: ParticipantTokenData,
@@ -988,6 +995,7 @@ export const Mutation = builder.mutationType({
               type: ElementOrderType,
               required: true,
             }),
+            availableFrom: t.arg({ type: 'Date', required: false }),
             resetTimeDays: t.arg.int({ required: true }),
           },
           resolve(_, args, ctx) {
@@ -1015,6 +1023,7 @@ export const Mutation = builder.mutationType({
               type: ElementOrderType,
               required: true,
             }),
+            availableFrom: t.arg({ type: 'Date', required: false }),
             resetTimeDays: t.arg.int({ required: true }),
           },
           resolve(_, args, ctx) {
@@ -1086,6 +1095,19 @@ export const Mutation = builder.mutationType({
           },
           resolve(_, args, ctx) {
             return MicroLearningService.publishMicroLearning(args, ctx)
+          },
+        }),
+
+      unpublishPracticeQuiz: t
+        .withAuth({ ...asUserWithCatalyst, ...asUserFullAccess })
+        .field({
+          nullable: true,
+          type: PracticeQuiz,
+          args: {
+            id: t.arg.string({ required: true }),
+          },
+          resolve(_, args, ctx) {
+            return PracticeQuizService.unpublishPracticeQuiz(args, ctx)
           },
         }),